Theme and User edit html, CSS, and javascript created
The html, CSS, and javascript for the theme setting and the user authentication setting were created. Template variables "profile" and "auth_form" created in "client/views.py". The profile model was added to the admin site. The Django form was created for the user authentication setting. Django widget tweaks was added.