Change Chats and Messages models
The Messages model was capitalized and the id and token fields were removed. In the Chats model, various fields were changed from text fields to other fields that make more sense. The verbose name attributes were added to all of the models and to the Chats model fields. Twenty Four migrations were made.